Update sgc.csv
Following discussion at WISOP meeting early December in WMO, and considering the interest of the Sensor Centre for Global Cache, I am proposing some changes to the metrics.
In particular, adding one with threshold on the delay to cache.
We need to agree on the threshold as well.
Suggestion here is 120s, 300s, 600s. Something not in the cache in 10 minutes is considered "missed".
I also suggest creating an additional metric for data not in ANY cache.